Discrete-Event-Simulation-of-Continuous-Review--Q-R--Inventory-Policy
A discrete event simulation is an approach to use a mathematical and/or logical representation of a physical system which encounters changes at discrete points in time. Discrete event simulation approach is based on imitation of the operations of a real world system. It also helps in generating an artificial history of a system and enables the user to visualize the system states through graphical means.

Discrete event simulation based approaches have several advantages such as:
1. Help study new designs without interrupting real systems.
2. Can help improve understanding of a system.
3. Provide decision frameworks that are realistic and robust.
4. Can help visualize and understand system dynamics.
5. Can help perform what-if kind of analysis for decision framework.
6. Uses actual system level logical or mathematical models and minimizes artificial assumptions or usage of stylized models.
7. Can be used for scenario analysis.
8. Can be used for dynamic decision making on a real time basis.

Discrete event simulation model a set of rules about the actual operations of a system. These rules can be:
- Algorithmic: A set of steps
- Mathematical: A set of equations defining some variable relationships pertaining to the underlying system being simulated, and
- Logical: A set of conditions that define the control of the system dynamics for simulation.